Reorganize overrides (#113)
* Move overrides for #blank? classes into their own files.

* Update windows testing to attempt to install the gem
