+++ lib/overrides/false_class.rb
@@ -1,0 +1,5 @@
+class FalseClass
+  def blank?
+    true
+  end
+end
