+++ lib/overrides/nil_class.rb
@@ -1,0 +1,5 @@
+class NilClass
+  def blank?
+    true
+  end
+end
