+++ lib/overrides/object.rb
@@ -1,0 +1,5 @@
+class Object
+  def blank?
+    respond_to?(:empty?) ? empty? : !self
+  end
+end
