+++ lib/overrides/string.rb
@@ -1,0 +1,5 @@
+class String
+  def blank?
+    self.strip.empty?
+  end
+end
