+++ lib/overrides/true_class.rb
@@ -1,0 +1,5 @@
+class TrueClass
+  def blank?
+    false
+  end
+end
